+++ CHANGELOG.md
@@ -23,7 +23,7 @@
 - Add multiple native `Event` fired on media actions (`play`, `pause`, etc.)
 - Add A11Y compatibility (`<button>`, `<input type="range">`, `aria-*`, `:focus-visible`)
 - Add the volume up/down shortcuts
-- Update DOM attributes from options and vice versa (`autoplay`, `playsinline`, `muted`, `loop`)
+- Update HTML attributes from options and vice versa (`autoplay`, `playsinline`, `muted`, `loop`)
 
 ### Updates